+++ rector.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 use Rector\CodeQuality\Rector\Equal\UseIdenticalOverEqualWithSameTypeRector;
 use Rector\Config\RectorConfig;
+use Rector\DeadCode\Rector\Expression\SimplifyMirrorAssignRector;
 use Rector\DeadCode\Rector\If_\RemoveAlwaysTrueIfConditionRector;
 use Rector\DeadCode\Rector\Plus\RemoveDeadZeroAndOneOperationRector;
 use Rector\Php84\Rector\MethodCall\NewMethodCallWithoutParenthesesRector;
@@ -24,5 +25,6 @@ return RectorConfig::configure()
         NewMethodCallWithoutParenthesesRector::class => [__DIR__ . '/tests'],
         RemoveDeadZeroAndOneOperationRector::class => [__DIR__ . '/tests'],
         UseIdenticalOverEqualWithSameTypeRector::class,
-        RemoveAlwaysTrueIfConditionRector::class => [__DIR__ . '/tests/BaseTestCase.php']
+        RemoveAlwaysTrueIfConditionRector::class => [__DIR__ . '/tests/BaseTestCase.php'],
+        SimplifyMirrorAssignRector::class
     ]);