+++ tests/Unit/TestCase.php
@@ -1,15 +1,6 @@
 <?php
 namespace Marcoconsiglio\ModularArithmetic\Tests\Unit;
 
-use Override;
-use Faker\Factory;
-use Faker\Generator;
-use PHPUnit\Framework\TestCase as PHPUnitTestCase;
-use Marcoconsiglio\ModularArithmetic\ModularInteger;
 use Marcoconsiglio\ModularArithmetic\Tests\BaseTestCase;
-use Marcoconsiglio\ModularArithmetic\Tests\Traits\WithFaker;
 
-class TestCase extends BaseTestCase
-{
- 
-}
+class TestCase extends BaseTestCase {}